Fix disableMultipart option of HTTPClientUpload
The disableMultipart option is already existed in 'HTTPClientUpload' class.

But, this option did not working in the 'HTTPClientUpload'.
So I fixed it.

And I removed the disableMultipart option on the FileItem. 
(it is revert of previous commit) (banksemi/ngx-upload@f5a41b68caf8be9b34ebfe7702b5098ed67f54fc)